[refactoring] 불필요한 생성자 삭제, 폴더 구조 변경
