+++ src/main/java/ssd/IO/OutputIO.java
@@ -3,8 +3,6 @@ package ssd.IO;
 import java.io.BufferedWriter;
 import java.io.FileWriter;
 
-import static ssd.SSDConstant.*;
-
 public class OutputIO extends IOHandler {
     public OutputIO(String path) {
         super(path);
